+++ attribution/views/perms.py
@@ -25,8 +25,7 @@
 ##############################################################################
 from django.core.exceptions import PermissionDenied
 
-from base.business.learning_units.perms import can_user_view_educational_information, \
-    can_user_edit_educational_information
+from attribution.business.perms import can_tutor_view_educational_information, can_user_edit_educational_information
 
 
 def tutor_can_edit_educational_information(view_func):
@@ -39,7 +38,7 @@ def tutor_can_edit_educational_information(view_func):
 
 def tutor_can_view_educational_information(view_func):
     def f_tutor_can_view_educational_information(request, learning_unit_year_id):
-        if not can_user_view_educational_information(request.user, learning_unit_year_id):
+        if not can_tutor_view_educational_information(request.user, learning_unit_year_id):
             raise PermissionDenied("User cannot view educational information")
         return view_func(request, learning_unit_year_id)
     return f_tutor_can_view_educational_information
